Add OpenStackPort
With this new resource OpenStackPort I propose some incremental changes
over the existing controller pattern proposed in OpenStackFlavor.

Functional/API changes:

* spec.id is the field for adoption. spec.resource.id will be the ID to
  be sent on resource creation, for the resources that allow that in
  OpenStack
* spec.unmanaged is unconditionally false by default. An adopted
  resource is not unmanaged by default any more. You can still set
  spec.unmanaged to true to prevent ORC from making changes against it.
* uses label key to point to a dependency when there can be many (label
  keys are unique). I used the MD5 hash of the name of the dependency to
  avoid hitting the 63-char mark.

Internal changes:
* Each reconcile loop now only applies once to kubeapi. Each loop either
  applies spec, or status.

---

In a previous iteration, this code had an issue where creation was run
twice. This is probably due to a race where the status change that
includes the openstackID of the newly created resource hasn't hit the
local object cache yet, when a new reconcile is triggered. This only
happened when the creation reconcile was triggered by a depedency that
becomes available (for example: create the OpenStackPort first, and then
the OpenStackNetwork it references).

This problem is mitigated within this PR when I introduced a 5-second
requeue when the dependency exists in Kubernetes but not in OpenStack.

Further mitigation might be provided when all resources (including
OpenStackNetwork in our example) adopt the rule of
one-apply-per-reconcile.

While we look for a proper fix, I propose to merge this given the
mitigations in place.