OpenStackImage: Use deterministic ID in OpenStack
The Image API in OpenStack allows to specify an ID upon image creation.
With this commit, the ID field of the image spec is guaranteed to be
populated before adoption (or creation) by generating an UUIDv5 based on
a constant UUID namespace and using the name of the Kubernetes resource.

Once the ID is guaranteed to be populated, the potential resource
already existing in OpenStack can be identified univocally.