Image, Flavor: Compute a unique UUID
IDs in OpenStack are also unique in time. Before this change, only the
name of the Kubernetes resource was used to generate a deterministic
OpenStack UUID. Which meant that once an image (or flavor) name was
taken on a given cloud, it could never be recreated.

With this patch, the creation time of the Kubernetes resource is also
used to compute a deterministic OpenStack UUID.