Generic finalizer handling
Move finalizer creation and deletion into generic methods. This ensures
that it is done consistently.

This also ensures that we don't add a finalizer until after we have an
authenticated openstack client. This prevents a situation when creating
objects with invalid credentials where the objects can't be deleted
without first fixing the credentials, which is unnecessary as they could
never have resulted in the creation of an OpenStack resource.