Docs: refresh getting started page
The old examples were only valid for the prototype. Update the example
to run with the current development version of ORC.

Also fix an issue with XDG_CONFIG_HOME variable being usually empty, and
remove the step of creating an OpenStackCloud object as it was only
a thing in the prototype.