Fix API for description
It turns out that the description field is not consistent across
OpenStack, and not even inside the same project.

As a consequence, drop the `OpenStackDescription` type, and introduce
more specialized types when it makes sense, for now, only the
`NeutronDescription` type.

All of the Neutron resources we have in ORC at the moment inherit from
the `standard_attr.HasStandardAttributes` class and have a description
sized to `DESCRIPTION_FIELD_SIZE` [1], equal to 255 bytes [2].

Flavor's description is an sqlalchemy.Text [3], and is limited to 65535
bytes.

[1] https://github.com/openstack/neutron-lib/blob/c4baf59/neutron_lib/db/standard_attr.py#L57
[2] https://github.com/openstack/neutron-lib/blob/fee4b23/neutron_lib/db/constants.py#L16
[3] https://github.com/openstack/nova/blob/031eda5/nova/db/api/models.py#L245