Improvements to Progressing condition
This change makes accurately reporting the Progressing condition easier
to understand and more consistent across controllers.

Several methods previously returned a slice of WaitingOnEvent.
WaitingOnEvent is renamed to ProgressStatus to reflect its use.
Returning a ProgressStatus indicates that some operation is incomplete
and we are expecting to be called again.

We pass this slice of ProgressStatus to the status writer. We no longer
pass a progressMessage, so ProgressStatus is now the only way to report
Progressing.

We update the logic setting Progressing so that it only considers any
returned error, and ProgressStatus.

Consequently, the GetCommonStatus method of ResourceStatusWriter is now
renamed to ResourceIsAvailable (and its signature updated), as it no
longer pretends to report Progressing.