Port: Make port available based on its status
Ports reports their statuses via a status field, we should take that
into consideration when computing if the port is available or not.

With this change, ports are reported as Available when the object has
a known ID for the resource, and the resource has ACTIVE or DOWN status.

Fixes #198