Deletion guard finalizers become responsibility of controllers
This change marks a fundmental change in how deletion guards operate.
Previously, the deletion guard controller was responsible for both
adding and removing the deletion guard controller. As it did not have
enough context to do better, it simply added the deletion guard
finalizer to all objects of the target type, i.e. every object of a type
which could be a dependency, whether or not it actually was.

With this change, the deletion guard is no longer responsible for adding
the finalizer; it is only responsible for removing it. Responsibility
now moves into the controllers which use the dependencies. i.e. With
this change, it is the responsibility of the port controller to ensure
that the network it references has its deletion guard finalizer. This
results in:
* Fewer finalizers being added
* Fewer reconcile events being generated by finalizers

To make it easier for controllers to do this robustly and consistently,
functionality for doing it is added to a new DeletionGuardDependency
object.

The former Dependency object is split in 2. Dependencies which do not
require a deletion guard because they do not need to block deletion of
their dependencies remain a Dependency. The rest use
DeletionGuardDependency instead. This adds a new GetDependencies method
which ensures that any returned dependency for the object being
reconciled already have the deletion guard finalizer.

We apply this pattern to all existing controllers. This is particularly
noisy in RouterInterface, as its implementation remains quite different
to most of the other controllers.