Fix deadlock deleting subnet with router interface
When a Subnet is created with a router reference, the subnet controller
automatically creates a RouterInterface object to represent the
attachment. This object is owned by the Subnet directly, not by the
actor who created the Subnet.

We only expect actors to delete their own objects. e.g. A user applying
and deleting manifests with kustomize:

$ kubectl create -k manifests
... and later
$ kubectl delete -k manifests

If these manifests created a Subnet, they will not reference the
RouterInterface which was created on the Subnet's behalf.

Previously this situation caused a deadlock:
* Subnet's DeleteResource does not run, because of the presence of the RouterInterface finalizer,
* which is never removed because the owning Subnet has not deleted it,
* because its DeleteResource did not run.

With this change, a deletion guard no longer blocks deletion of an
object which owns its dependency. The assumption here is that the owner
is explicitly aware of the objects it created, and will act correctly.