CI: Make resource name consistent with other tests
We want the resource name to always start with `<test-name>`, with
potentially a suffix when there is more than one resource and we need to
differentiate them. This consistency allows to mostly prevent name
clashes and programming mistakes in the asserts when copy-pasting from
other tests.

This fixes a few mistakes where we were not checking the right resource
deletion.