Fix import of server where name is not specified
Commit 3deb8ee347fe1ded293d96e69ed941768431ba30 fixed import for names
that are substring of another server name, however this broke the
import for when we don't specify a name in the filter. In which case,
the filter would look for servers that don't have a name. This is
obviously not what we want.

This broke the `import-error` test and we didn't notice because we had
forgotten to enable the suite as part of the CI runs.

Now, only filter by name when a name is set in the filter.