build: Use make to build binary in container image
Introduce a new `build-manager` target that does not have dependency on
`generate` as it proved problematic while building the docker image.
We're using this new target to create a binary containing the build-time
variables.

Also copy the whole directory to the builder stage so that we have the
git context at build time.