Add weekly security scan using govulncheck and trivy
This adds a new make target "verify-security" that runs govulncheck agains the code base
and also builds the contaier image and scans it with trivy.
I tried keeping things similar to how they are done elsewhere in the code.
This is all based on how they do it in Cluster API (hence the copyright text in the scripts).

Probably the most controversial thing here is that I added the GO_VERSION variable. The idea is to make sure we know what exact version is used.

Signed-off-by: Lennart Jern <lennart.jern@est.tech>