Makefile: introduce a generate-bundle target
For new controllers, we need to re-generate the OLM bundle manifest so
that it knows about the new CRDs. We now have a dedicated
`generate-bundle` target for it. We could it a dependency of the
`generate` target so that it's always up-to-date and checked in CI via
the `verify-generated` target, however I think it's a bad idea because
it would mean installing the operator-sdk in every CI run which is
costly for no real gain: the manifest is always generated (and thus
up-to-date) when building the OLM bundle.