Refresh volumes after an attachment event
Trigger a refresh of the volumes objects when they're being attached or
detached from a server, so that their status more closely reflect
reality.

This introduces a new `attachments` field in the object spec, that is
modified whenever a server attaches of detaches a volume. Updating the
spec triggers a new reconcile of the object, which status is then
updated. We treat the refresh of the attachment as "best-effort" and
will not fail the reconcile if the field couldn't be updated.

This is the solution highlighted in #282, however we couldn't implement
a similar fix for ports just yet as the server controller does not
reconcile ports. We'll do it in a follow-up change, using the same
technique.