Bump kube-api-linter
Enable all of the new linters except for `optionalfields` which would
require us to change many of the API fields to pointer type. We're not
yet ready to do so as this would likely require a new major bump.

Also add new setting for the `requiredfields` linter who got extra
capability and now flags more errors. Still, we're getting errors which
we can fix in a backward compatible way in a next commit.