Bump kuttl timeout
We started seeing the CI flake on the server-update test, where the port
does not have the expected status. This is most likely a timeout occurring
in CI, as this is both our longest test to run, and the port check is
a new addition that makes the test take even longer.

Let's give kuttl more time to run the tests and see if it stabilizes the
CI.